filtering layout improvements
