+++ api/filtering_common.py
@@ -1,0 +1,32 @@
+from enum import Enum
+
+
+class OPERATION_SETS(Enum):
+    eq = ["eq", "contains"]  # add contains for when it's a list
+    matches = ["matches"]
+    is_op = ["is"]  # "is" is reserved
+    range = ["lt", "gt", "lte", "gte"]
+
+
+SPEC_OPERATIONS_LOOKUP = {
+    "string": OPERATION_SETS.eq.value,
+    "wildcard": OPERATION_SETS.eq.value,  # because on our side we want eq
+    "boolean": OPERATION_SETS.eq.value,
+    "range": OPERATION_SETS.range.value,
+    "operating_system": OPERATION_SETS.range.value,
+}
+
+GRAPHQL_OPERATIONS_LOOKUP = {
+    "string": OPERATION_SETS.eq.value[0],
+    "wildcard": OPERATION_SETS.eq.value[0],  # because on our side we want eq
+    "boolean": OPERATION_SETS.is_op.value[0],
+    "range": OPERATION_SETS.eq.value[0],
+}
+
+
+def lookup_operations(filter_type):
+    return SPEC_OPERATIONS_LOOKUP[filter_type]
+
+
+def lookup_graphql_operations(filter_type):
+    return GRAPHQL_OPERATIONS_LOOKUP[filter_type]
